+++ CHANGELOG.md
@@ -257,6 +257,7 @@ Please submit any feedback you may have for this feature at https://github.com/h
 - console: Hide TimescaleDB internal schema from data tab
 - console: support naming convention in source customization for postgres DB [CON-297]
 - console: fix bug where analyze button would stay in analyzing state after analyze failed
+- console: fix broken table relationship query on non-default schema
 - build: changes to the `hasura/graphql-engine` docker image:
   - Debian graphql-engine docker images (`hasura/graphql-engine:<VERSION>`) now use a ubuntu base instead of debian.
   - Debian flavour of images (`hasura/graphql-engine:<VERSION>.debian`) are still published to Docker Hub.